backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart==0.0.12
openai>=1.0.0
httpx>=0.27.0
snowflake-connector-python>=3.6.0
pgvector>=0.3.0
Pillow>=10.2.0
